createGroup() - GroupController - and saveGroup() - saveGroup() refactored with new exceptions
