+++ src/main/java/com/kodilla/ecommercee/exception/groupException/GroupNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.exception;
+package com.kodilla.ecommercee.exception.groupException;
 
 public class GroupNotFoundException extends Exception {
 }
