+++ src/main/java/com/kodilla/ecommercee/repository/GroupRepository.java
@@ -1,11 +1,16 @@
 package com.kodilla.ecommercee.repository;
 
 import com.kodilla.ecommercee.domain.Group;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
 public interface GroupRepository extends CrudRepository<Group, Long> {
+    @Override
     List<Group> findAll();
+
+    @Query
+    boolean existsGroupByGroupName(String groupName);
 }
